[GDE] Modification of runnable set_motor_angle

+++ Servo_Motor_Driver_set_motor_angle.c
@@ -1,7 +1,7 @@
 #include "Fmi_Cfg.h" // Ids for <XXX>WriteToSharedObj
 #include "ASimSharedObjUtils.h" // <XXX>WriteToSharedObj
 
-void set_motor_angle(float angle){
+void set_motor_angle(uint16_t angle){
     // FIXME
 	(void) angle;
 }
